+++ src/main/java/com/tasksprints/auction/domain/auction/repository/AuctionRepository.java
@@ -3,6 +3,7 @@ package com.tasksprints.auction.domain.auction.repository;
 
 import com.tasksprints.auction.domain.auction.model.Auction;
 import com.tasksprints.auction.domain.auction.model.AuctionCategory;
+import com.tasksprints.auction.domain.auction.repository.support.AuctionCriteriaRepository;
 import com.tasksprints.auction.domain.product.model.ProductCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,7 +12,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface AuctionRepository extends JpaRepository<Auction, Long>,AuctionCriteriaRepository{
+public interface AuctionRepository extends JpaRepository<Auction, Long>, AuctionCriteriaRepository {
 
 
     @Query("SELECT a FROM auction a WHERE a.seller.id = :userId")
